# Conflicts:
#	Cargo.lock
#	Cargo.toml
#	genesis/generated/halley/genesis
#	genesis/generated/proxima/genesis
#	kube/manifest/starcoin-proxima.yaml
#	vm/stdlib/compiled/latest/error_descriptions/error_descriptions.errmap
#	vm/stdlib/compiled/latest/stdlib/011_Option.mv
#	vm/stdlib/compiled/latest/stdlib/012_Event.mv
#	vm/stdlib/compiled/latest/stdlib/013_BCS.mv
#	vm/stdlib/compiled/latest/stdlib/013_Event.mv
#	vm/stdlib/compiled/latest/stdlib/014_CoreAddresses.mv
#	vm/stdlib/compiled/latest/stdlib/014_Event.mv
#	vm/stdlib/compiled/latest/stdlib/015_CoreAddresses.mv
#	vm/stdlib/compiled/latest/stdlib/015_Timestamp.mv
#	vm/stdlib/compiled/latest/stdlib/016_CoreAddresses.mv
#	vm/stdlib/compiled/latest/stdlib/016_Option.mv
#	vm/stdlib/compiled/latest/stdlib/016_Timestamp.mv
#	vm/stdlib/compiled/latest/stdlib/017_Timestamp.mv
#	vm/stdlib/compiled/latest/stdlib/018_Option.mv
#	vm/stdlib/compiled/latest/stdlib/035_Account.mv
#	vm/stdlib/compiled/latest/stdlib/036_AccountScripts.mv
#	vm/stdlib/compiled/latest/stdlib/038_TreasuryWithdrawDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/039_BlockReward.mv
#	vm/stdlib/compiled/latest/stdlib/040_Block.mv
#	vm/stdlib/compiled/latest/stdlib/040_Collection.mv
#	vm/stdlib/compiled/latest/stdlib/040_TreasuryWithdrawDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/041_BlockReward.mv
#	vm/stdlib/compiled/latest/stdlib/041_Collection2.mv
#	vm/stdlib/compiled/latest/stdlib/041_TreasuryWithdrawDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/042_BlockReward.mv
#	vm/stdlib/compiled/latest/stdlib/042_Collection.mv
#	vm/stdlib/compiled/latest/stdlib/043_Collection.mv
#	vm/stdlib/compiled/latest/stdlib/043_Collection2.mv
#	vm/stdlib/compiled/latest/stdlib/043_ConsensusStrategy.mv
#	vm/stdlib/compiled/latest/stdlib/044_Collection2.mv
#	vm/stdlib/compiled/latest/stdlib/044_DaoVoteScripts.mv
#	vm/stdlib/compiled/latest/stdlib/045_Compare.mv
#	vm/stdlib/compiled/latest/stdlib/045_ConsensusStrategy.mv
#	vm/stdlib/compiled/latest/stdlib/045_DummyToken.mv
#	vm/stdlib/compiled/latest/stdlib/046_ConsensusStrategy.mv
#	vm/stdlib/compiled/latest/stdlib/046_DaoVoteScripts.mv
#	vm/stdlib/compiled/latest/stdlib/046_DummyTokenScripts.mv
#	vm/stdlib/compiled/latest/stdlib/047_DaoVoteScripts.mv
#	vm/stdlib/compiled/latest/stdlib/047_DummyToken.mv
#	vm/stdlib/compiled/latest/stdlib/048_DummyToken.mv
#	vm/stdlib/compiled/latest/stdlib/048_DummyTokenScripts.mv
#	vm/stdlib/compiled/latest/stdlib/048_Epoch.mv
#	vm/stdlib/compiled/latest/stdlib/049_DummyTokenScripts.mv
#	vm/stdlib/compiled/latest/stdlib/049_FixedPoint32.mv
#	vm/stdlib/compiled/latest/stdlib/050_EVMAddress.mv
#	vm/stdlib/compiled/latest/stdlib/050_Epoch.mv
#	vm/stdlib/compiled/latest/stdlib/050_GenesisSignerCapability.mv
#	vm/stdlib/compiled/latest/stdlib/051_Epoch.mv
#	vm/stdlib/compiled/latest/stdlib/051_Oracle.mv
#	vm/stdlib/compiled/latest/stdlib/052_FixedPoint32.mv
#	vm/stdlib/compiled/latest/stdlib/053_FixedPoint32.mv
#	vm/stdlib/compiled/latest/stdlib/053_GenesisSignerCapability.mv
#	vm/stdlib/compiled/latest/stdlib/053_STCUSDOracle.mv
#	vm/stdlib/compiled/latest/stdlib/054_GenesisSignerCapability.mv
#	vm/stdlib/compiled/latest/stdlib/054_Offer.mv
#	vm/stdlib/compiled/latest/stdlib/054_Oracle.mv
#	vm/stdlib/compiled/latest/stdlib/055_NFT.mv
#	vm/stdlib/compiled/latest/stdlib/055_Oracle.mv
#	vm/stdlib/compiled/latest/stdlib/056_LanguageVersion.mv
#	vm/stdlib/compiled/latest/stdlib/056_PriceOracle.mv
#	vm/stdlib/compiled/latest/stdlib/056_STCUSDOracle.mv
#	vm/stdlib/compiled/latest/stdlib/057_MerkleProof.mv
#	vm/stdlib/compiled/latest/stdlib/057_Offer.mv
#	vm/stdlib/compiled/latest/stdlib/057_STCUSDOracle.mv
#	vm/stdlib/compiled/latest/stdlib/058_MerkleNFTDistributor.mv
#	vm/stdlib/compiled/latest/stdlib/058_NFT.mv
#	vm/stdlib/compiled/latest/stdlib/058_Offer.mv
#	vm/stdlib/compiled/latest/stdlib/059_LanguageVersion.mv
#	vm/stdlib/compiled/latest/stdlib/059_NFT.mv
#	vm/stdlib/compiled/latest/stdlib/060_LanguageVersion.mv
#	vm/stdlib/compiled/latest/stdlib/060_MerkleProof.mv
#	vm/stdlib/compiled/latest/stdlib/061_MerkleNFTDistributor.mv
#	vm/stdlib/compiled/latest/stdlib/061_MerkleProof.mv
#	vm/stdlib/compiled/latest/stdlib/062_Genesis.mv
#	vm/stdlib/compiled/latest/stdlib/062_MerkleNFTDistributor.mv
#	vm/stdlib/compiled/latest/stdlib/064_GenesisNFT.mv
#	vm/stdlib/compiled/latest/stdlib/065_Genesis.mv
#	vm/stdlib/compiled/latest/stdlib/065_MintDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/065_StdlibUpgradeScripts.mv
#	vm/stdlib/compiled/latest/stdlib/066_Genesis.mv
#	vm/stdlib/compiled/latest/stdlib/067_GenesisNFTScripts.mv
#	vm/stdlib/compiled/latest/stdlib/068_IdentifierNFTScripts.mv
#	vm/stdlib/compiled/latest/stdlib/068_MintDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/069_MintDaoProposal.mv
#	vm/stdlib/compiled/latest/stdlib/069_OnChainConfigScripts.mv
#	vm/stdlib/compiled/latest/stdlib/070_ModuleUpgradeScripts.mv
#	vm/stdlib/compiled/latest/stdlib/070_PriceOracleAggregator.mv
#	vm/stdlib/compiled/latest/stdlib/071_NFTGallery.mv
#	vm/stdlib/compiled/latest/stdlib/072_OnChainConfigScripts.mv
#	vm/stdlib/compiled/latest/stdlib/073_OnChainConfigScripts.mv
#	vm/stdlib/compiled/latest/stdlib/073_PriceOracleAggregator.mv
#	vm/stdlib/compiled/latest/stdlib/073_SharedEd25519PublicKey.mv
#	vm/stdlib/compiled/latest/stdlib/074_PriceOracleAggregator.mv
#	vm/stdlib/compiled/latest/stdlib/074_TransactionTimeout.mv
#	vm/stdlib/compiled/latest/stdlib/075_PriceOracleScripts.mv
#	vm/stdlib/compiled/latest/stdlib/075_TransactionManager.mv
#	vm/stdlib/compiled/latest/stdlib/076_SharedEd25519PublicKey.mv
#	vm/stdlib/compiled/latest/stdlib/076_TransferScripts.mv
#	vm/stdlib/compiled/latest/stdlib/077_Signature.mv
#	vm/stdlib/compiled/latest/stdlib/077_TreasuryScripts.mv
#	vm/stdlib/compiled/latest/stdlib/078_SharedEd25519PublicKey.mv
#	vm/stdlib/compiled/latest/stdlib/080_TransactionTimeout.mv
#	vm/stdlib/compiled/latest/stdlib/080_YieldFarming.mv
#	vm/stdlib/compiled/latest/stdlib/081_TransactionManager.mv
#	vm/stdlib/compiled/latest/stdlib/081_YieldFarmingV2.mv
#	vm/stdlib/compiled/latest/stdlib/082_TransactionTimeout.mv
#	vm/stdlib/compiled/latest/stdlib/082_TransferScripts.mv
#	vm/stdlib/compiled/latest/stdlib/083_TransactionManager.mv
#	vm/stdlib/compiled/latest/stdlib/083_TreasuryScripts.mv
#	vm/stdlib/compiled/latest/stdlib/084_TransferScripts.mv
#	vm/stdlib/compiled/latest/stdlib/085_TreasuryScripts.mv
#	vm/stdlib/compiled/latest/stdlib/086_YieldFarming.mv
#	vm/stdlib/compiled/latest/stdlib/087_U256.mv
#	vm/stdlib/compiled/latest/stdlib/087_YieldFarmingV2.mv
#	vm/stdlib/compiled/latest/stdlib/088_YieldFarming.mv
#	vm/stdlib/compiled/latest/stdlib/089_YieldFarmingV2.mv

+++ commons/accumulator/src/node_index.rs
@@ -80,12 +80,12 @@ impl NodeIndex {
 
     pub fn root_from_leaf_count(leaf_count: LeafCount) -> Self {
         assert!(leaf_count > 0);
-        Self::root_from_leaf_index((leaf_count - 1) as u64)
+        Self::root_from_leaf_index(leaf_count - 1)
     }
 
     pub fn root_level_from_leaf_count(leaf_count: LeafCount) -> u32 {
         assert!(leaf_count > 0);
-        let index = (leaf_count - 1) as u64;
+        let index = leaf_count - 1;
         MAX_ACCUMULATOR_PROOF_DEPTH as u32 + 1 - index.leading_zeros()
     }
 
@@ -350,7 +350,7 @@ impl Iterator for FrozenSubtreeSiblingIterator {
         self.remaining_new_leaves -= next_subtree_leaves;
 
         Some(NodeIndex::from_inorder_index(
-            (first_leaf_index + last_leaf_index) as u64,
+            first_leaf_index + last_leaf_index,
         ))
     }
 }
